requirements.txt
plotly
streamlit
mysql.connector
